Revert "rake refactor: simplify cleaning"
This reverts commit e920e5ad6a1fc9ec1b919595f8f4e0439d63b9b5.

+++ scripts/tools/shell.rb
@@ -32,12 +32,12 @@ module Shell
 
   def self.rm_rf(dir)
     puts "[rm_rf] #{Shell.cwd}> #{dir}" if Shell.is_verbose_hidden
-    FileUtils.rm_rf(dir)
+    FileUtils.rm_rf(dir) if File.exist?(dir)
   end
 
   def self.rm(file)
     puts "[rm   ] #{Shell.cwd}> #{file}" if Shell.is_verbose_hidden
-    FileUtils.rm_f(file)
+    FileUtils.rm(file) if File.exist?(file)
   end
 
   def self.chdir(dir, &block)
